+++ apps/v15/src/app/force-cache/_get-hello.ts
@@ -1,0 +1,11 @@
+export const getHello = async () => {
+  const res = await fetch("http://localhost:3010/api/hello", {
+    cache: "force-cache",
+  });
+  if (!res.ok) {
+    throw new Error("Failed to fetch data");
+  }
+  const result = res.json();
+
+  return result;
+};
