+++ apps/v15/src/app/force-cache/_hello-button.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+import { getHelloAction } from "./action";
+
+export function HelloButton() {
+  return (
+    <button
+      onClick={async () => {
+        const result = await getHelloAction();
+        console.log(result);
+      }}
+    >
+      Hello Button
+    </button>
+  );
+}
