+++ apps/v15/src/app/force-cache/_hello.tsx
@@ -1,16 +1,10 @@
-const getHello = async () => {
-  const res = await fetch("http://localhost:3010/api/hello", {
-    cache: "force-cache",
-  });
-  if (!res.ok) {
-    throw new Error("Failed to fetch data");
-  }
-  const result = res.json();
-
-  return result;
-};
+import { getHello } from "./_get-hello";
 
 export async function Hello() {
   const result = await getHello();
-  return <pre>{JSON.stringify(result, null, 2)}</pre>;
+  return (
+    <div>
+      <pre>{JSON.stringify(result, null, 2)}</pre>
+    </div>
+  );
 }
