+++ apps/v15/src/app/force-cache/action.ts
@@ -1,0 +1,7 @@
+"use server";
+
+import { getHello } from "./_get-hello";
+
+export const getHelloAction = async () => {
+  return await getHello();
+};
