+++ apps/v15/src/app/force-cache/page.tsx
@@ -1,8 +1,10 @@
 import { Hello } from "./_hello";
+import { HelloButton } from "./_hello-button";
 
-export default async function Page() {
+export default function Page() {
   return (
     <main>
+      <HelloButton />
       <Hello />
       <Hello />
       <Hello />
